pagination unit tests added

+++ api/internal/models/databases.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Database struct {
-	ID          uint           `json:"id" gorm:"primaryKey;autoIncrement;unique"`
+	ID          uint64         `json:"id" gorm:"primaryKey;autoIncrement;unique"`
 	Name        string         `json:"name" gorm:"unique" validate:"required"`
 	Description *string        `json:"description"`
 	CreatedAt   time.Time      `json:"createdAt"`
@@ -18,7 +18,7 @@ type Database struct {
 	DeletedAt   gorm.DeletedAt `json:"deletedAt" gorm:"index"`
 }
 
-func (d Database) GetID() uint { return d.ID }
+func (d Database) GetID() uint64 { return d.ID }
 
 type DatabaseResponse struct {
 	Data []Database
